refactor: split yaml process system
