+++ src/libecalc/presentation/yaml/yaml_types/process/yaml_stream_distribution.py
@@ -1,0 +1,36 @@
+from pydantic import Field
+from typing import Literal, Annotated
+
+from libecalc.presentation.yaml.yaml_types import YamlBase
+from libecalc.presentation.yaml.yaml_types.components.yaml_expression_type import YamlExpressionType
+from libecalc.presentation.yaml.yaml_types.process.yaml_process_pipeline import ProcessPipelineReference
+from libecalc.presentation.yaml.yaml_types.streams.yaml_inlet_stream import YamlInletStream
+
+
+StreamRef = str
+
+
+class YamlOverflow(YamlBase):
+    from_reference: ProcessPipelineReference
+    to_reference: ProcessPipelineReference
+
+
+class YamlCommonStreamSetting(YamlBase):
+    rate_fractions: list[YamlExpressionType]
+    overflow: list[YamlOverflow] | None = None
+
+
+class YamlCommonStreamDistribution(YamlBase):
+    method: Literal["COMMON_STREAM"]
+    inlet_stream: StreamRef | YamlInletStream
+    settings: list[YamlCommonStreamSetting]
+
+
+class YamlIndividualStreamDistribution(YamlBase):
+    method: Literal["INDIVIDUAL_STREAMS"]
+    inlet_streams: list[StreamRef | YamlInletStream]
+
+
+YamlStreamDistribution = Annotated[
+    YamlCommonStreamDistribution | YamlIndividualStreamDistribution, Field(discriminator="method")
+]
