+++ tests/libecalc/input/mappers/test_model_mapper.py
@@ -8,11 +8,6 @@ from libecalc.presentation.yaml.domain.reference_service import ReferenceService
 from libecalc.presentation.yaml.mappers.consumer_function_mapper import CompressorModelMapper
 from libecalc.presentation.yaml.mappers.yaml_path import YamlPath
 from libecalc.presentation.yaml.yaml_entities import MemoryResource
-from libecalc.presentation.yaml.yaml_types.components.yaml_process_system import (
-    YamlCompressor,
-    YamlCompressorStageProcessSystem,
-    YamlProcessPipeline,
-)
 from libecalc.presentation.yaml.yaml_types.facility_model.yaml_facility_model import (
     YamlGeneratorSetModel,
     YamlPumpChartSingleSpeed,
@@ -20,6 +15,11 @@ from libecalc.presentation.yaml.yaml_types.facility_model.yaml_facility_model im
     YamlTabularModel,
 )
 from libecalc.presentation.yaml.yaml_types.models import YamlCompressorChart, YamlFluidModel, YamlTurbine
+from libecalc.presentation.yaml.yaml_types.process.yaml_process_pipeline import (
+    YamlCompressorStageProcessSystem,
+    YamlProcessPipeline,
+)
+from libecalc.presentation.yaml.yaml_types.process.yaml_process_units import YamlCompressor
 from libecalc.presentation.yaml.yaml_types.streams.yaml_inlet_stream import YamlInletStream
